fixed "E=10000" transaltion issue

+++ org.eclipse.iee.translator.antlr/src/org/eclipse/iee/translator/antlr/translator/TypeVisitior.java
@@ -30,7 +30,7 @@ public class TypeVisitior extends MathBaseVisitor<VariableType> {
 	
 	@Override
 	public VariableType visitIntNumber(IntNumberContext ctx) {
-		return VariableType.INT;
+		return VariableType.DOUBLE;
 	}
 	
 	@Override
